+++ components/01-atoms/LoadingIndicator.tsx
@@ -1,7 +1,5 @@
 import React from "react";
 
-const LoadingIndicator = () => (
+export const LoadingIndicator = () => (
   <div className="animate-spin rounded-full h-4 w-4 border-t-2 border-white"></div>
 );
-
-export default LoadingIndicator;
