+++ components/01-atoms/SearchBar.tsx
@@ -1,11 +1,11 @@
 import cc from "classcat";
 import { useContext, useEffect, useState } from "react";
+import { MagnifyingGlassIcon } from "@/components/01-atoms/icons";
 import {
-  MagnifyingGlassIcon,
   SelectAuthedUserChain,
   SelectDestinyChain,
   SwapContext,
-} from ".";
+} from "@/components/01-atoms";
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
 import { ENS } from "web3-eth-ens";
 import Web3 from "web3";
