+++ components/01-atoms/SelectDestinyChain.tsx
@@ -1,6 +1,7 @@
-import { Dialog, Transition } from "@headlessui/react";
 import { Fragment, useContext, useState } from "react";
-import { EthereumIcon, PolygonIcon, SwapContext } from ".";
+import { Dialog, Transition } from "@headlessui/react";
+import { SwapContext } from "@/components/01-atoms";
+import { EthereumIcon, PolygonIcon } from "@/components/01-atoms/icons";
 import { ChainInfo, SupportedNetworks } from "@/lib/client/constants";
 import cc from "classcat";
