+++ components/01-atoms/icons/index.tsx
@@ -1,0 +1,10 @@
+export * from "./DangerIcon";
+export * from "./EthereumIcon";
+export * from "./MagnifyingGlassIcon";
+export * from "./PaperPlane";
+export * from "./PersonIcon";
+export * from "./PolygonIcon";
+export * from "./SelectUserIcon";
+export * from "./SwapIcon";
+export * from "./SwaplaceIcon";
+export * from "./WalletIcon";
