+++ components/01-atoms/index.ts
@@ -1,15 +1,9 @@
 export * from "./ConnectWallet";
-export * from "../02-molecules/NftCard";
 export * from "./SearchBar";
 export * from "./SelectAuthedUserChain";
 export * from "./Tab";
 export * from "./SwapContext";
-export * from "./icons/SwaplaceIcon";
-export * from "./icons/MagnifyingGlassIcon";
-export * from "./icons/SwapIcon";
-export * from "./icons/PaperPlane";
 export * from "./SelectDestinyChain";
-export * from "./icons/EthereumIcon";
-export * from "./icons/PolygonIcon";
 export * from "./ConfirmSwapModal";
 export * from "./TransactionResultModal";
+export * from "./LoadingIndicator";
