+++ components/02-molecules/index.tsx
@@ -1,6 +1,6 @@
 export * from "./CardHome";
-export * from "./TheHeader";
-export * from "../03-organisms/NftsShelf";
+export * from "./NftCard";
 export * from "./NftsList";
-export * from "./SwapStation";
 export * from "./OfferSummary";
+export * from "./SwapStation";
+export * from "./TheHeader";
