+++ components/03-organisms/NftsShelf.tsx
@@ -1,18 +1,13 @@
 import { useContext, useEffect, useState } from "react";
-import {
-  NFT,
-  ChainInfo,
-  NFTsQueryStatus,
-  ADDRESS_ZERO,
-} from "@/lib/client/constants";
+import { NFT, ChainInfo, NFTsQueryStatus } from "@/lib/client/constants";
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
-import { NftsList } from "../02-molecules";
-import { getNftsFrom } from "@/lib/client/blockchain-data";
-import { SwapContext, SwapIcon } from "../01-atoms";
-import { EthereumAddress } from "@/lib/shared/types";
 import { useNetwork } from "wagmi";
-import { SelectUserIcon } from "../01-atoms/icons/SelectUserIcon";
+import { getNftsFrom } from "@/lib/client/blockchain-data";
 import { useTheme } from "next-themes";
+import { EthereumAddress } from "@/lib/shared/types";
+import { SwapContext } from "@/components/01-atoms";
+import { SelectUserIcon } from "@/components/01-atoms/icons";
+import { NftsList } from "@/components/02-molecules";
 
 /**
  *
