+++ components/03-organisms/index.tsx
@@ -1,1 +1,2 @@
+export * from "./NftsShelf";
 export * from "./SwappingShelfs";
