+++ components/04-templates/SwapSection.tsx
@@ -1,6 +1,6 @@
 import { SearchBar } from "@/components/01-atoms";
-import { SwappingShelfs } from "@/components/03-organisms";
 import { SwapStation } from "@/components/02-molecules";
+import { SwappingShelfs } from "@/components/03-organisms";
 
 export const SwapSection = () => {
   return (
